backend/requirements.txt
flask
flask-cors
pymongo
spacy
flask-bcrypt
pyjwt
